REFACTOR RiskInterval to be a nested class in DistributionOfRisk
ADD jackson deserialiser annotation on DistributionOfRisk class
