+++ src/main/java/no/oslomet/aaas/analyzer/ARXAnalyzer.java
@@ -3,7 +3,6 @@ package no.oslomet.aaas.analyzer;
 import no.oslomet.aaas.model.AnalyzeResult;
 import no.oslomet.aaas.model.DistributionOfRisk;
 import no.oslomet.aaas.model.Request;
-import no.oslomet.aaas.model.RiskInterval;
 import no.oslomet.aaas.utils.ARXPayloadAnalyser;
 import no.oslomet.aaas.utils.DataFactory;
 import org.deidentifier.arx.ARXPopulationModel;
@@ -12,7 +11,6 @@ import org.deidentifier.arx.DataHandle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -34,8 +32,7 @@ public class ARXAnalyzer implements Analyzer {
         DataHandle dataToAnalyse = data.getHandle();
         ARXPopulationModel pModel= ARXPopulationModel.create(data.getHandle().getNumRows(), 0.01d);
         Map<String,String> analysisMetrics = ARXPayloadAnalyser.getPayloadAnalyzeData(dataToAnalyse,pModel);
-        List<RiskInterval> listRiskInterval = distributionOfRisk(dataToAnalyse,pModel).getRiskIntervalList();
-        return new AnalyzeResult(analysisMetrics,listRiskInterval);
+        return new AnalyzeResult(analysisMetrics,distributionOfRisk(dataToAnalyse,pModel));
     }
 
     private DistributionOfRisk distributionOfRisk(DataHandle dataToAnalyse, ARXPopulationModel pModel){
