+++ src/main/java/no/oslomet/aaas/model/AnalyzeResult.java
@@ -1,6 +1,5 @@
 package no.oslomet.aaas.model;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -10,7 +9,7 @@ public class AnalyzeResult {
 
     private final Map<String, String> metrics;
 
-    private final List<RiskInterval> distributionOfRisk;
+    private final DistributionOfRisk distributionOfRisk;
 
     /***
      * Setter method for analysis result metrics. Fills the object with the results from the analysis against
@@ -18,7 +17,7 @@ public class AnalyzeResult {
      * @param metrics HashMap containing a String of fields that describe the type of risk and a
      *                String containing its risk value
      */
-    public AnalyzeResult(Map<String, String> metrics, List<RiskInterval> distributionOfRisk ) {
+    public AnalyzeResult(Map<String, String> metrics, DistributionOfRisk distributionOfRisk ) {
         this.metrics = metrics;
         this.distributionOfRisk = distributionOfRisk;
     }
@@ -33,7 +32,7 @@ public class AnalyzeResult {
     }
 
 
-    public List<RiskInterval> getDistributionOfRisk() {
+    public DistributionOfRisk getDistributionOfRisk() {
         return distributionOfRisk;
     }
 }
